+++ lua/obsidian/autocmds.lua
@@ -59,13 +59,6 @@ local function bufenter_callback(ev)
     end, { buffer = true, desc = "Obsidian Previous Link" })
   end
 
-  -- Inject completion sources, providers to their plugin configurations
-  if opts.completion.nvim_cmp then
-    require("obsidian.completion.plugin_initializers.nvim_cmp").inject_sources()
-  elseif opts.completion.blink then
-    require("obsidian.completion.plugin_initializers.blink").inject_sources()
-  end
-
   require("obsidian.lsp").start(ev.buf)
 
   if opts.footer.enabled then