+++ lua/obsidian/commands/new_from_template.lua
@@ -3,6 +3,6 @@ return function(data)
   local id = table.concat(data.fargs, " ", 1, #data.fargs - 1)
   local template = data.fargs[#data.fargs]
   require("obsidian.actions").new_from_template(id, template, function(note)
-    note:open()
+    note:open { sync = true }
   end)
 end